Login and Registration
This is my attempt at making the login and registration framework for the project.  It generally follows the video I sent in discord, but can be manipulated for our needs.  It is based on API controllers and models, not MVC so idk if that's what we're going for, but we can figure that out tomorrow.