HTML to PDF generator
GeneratePDF() in the HtmlToPDFController.cs file is meant to take in an html string, convert it to a pdf and then return a file that is of type pdf and named ReportCard.pdf.  I was unable to test this because the sample html being pulled from a page wouldn't load so hope this works.